requirements.txt
pyserial 
